Fixing script compiling bug in linux (only eclipse).
This bug happened only if you started solys from eclipse. In the
.classpath file of the scripts plugin, we remove a leading '/' for the
windows version, e.g. '/C:/path/to/lib.jar' becomes to
'C:/path/to/lib.jar'. In *-ux based systems this leads sometimes to
wrong paths, e.g. '/home/...' to 'home/...'.

Solution is to remove the '/' only if on Windows.